Rename feature modules to lowercase for PEP 8 compliance
Rename PWM.py, DeepLC.py, netMHCpan.py, and netMHCIIpan.py to their
lowercase equivalents (pwm.py, deeplc.py, netmhcpan.py, netmhciipan.py).
Update all import statements in feature_generation.py and module paths
in docs/api/features.md accordingly.
